requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
arrow==1.4.0
arxiv==3.0.0
attrs==26.1.0
aws-sam-cli==1.159.1
aws-sam-translator==1.109.0
aws_lambda_builders==1.64.0
awscrt==0.31.2
binaryornot==0.6.0
blinker==1.9.0
boto3==1.42.70
boto3-stubs==1.42.97
botocore==1.42.97
botocore-stubs==1.42.41
certifi==2026.4.22
cffi==2.0.0
cfn-lint==1.48.1
charset-normalizer==3.4.7
chevron==0.14.0
click==8.1.8
cohere==6.1.0
colorama==0.4.6
cookiecutter==2.7.1
cryptography==46.0.7
dataclasses-json==0.6.7
datasets==4.8.5
dateparser==1.4.0
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.18.0
fastavro==1.12.2
feedparser==6.0.12
filelock==3.29.0
Flask==3.1.3
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.5.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.12.2
idna==3.13
iniconfig==2.3.0
instructor==1.15.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.15
langchain-aws==1.4.5
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.2
langchain-ollama==1.1.0
langchain-openai==1.2.1
langchain-protocol==0.0.13
langchain-text-splitters==1.1.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.12
langgraph-sdk==0.3.13
langsmith==0.7.38
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy-boto3-apigateway==1.42.68
mypy-boto3-cloudformation==1.42.3
mypy-boto3-ecr==1.42.86
mypy-boto3-iam==1.42.64
mypy-boto3-kinesis==1.42.41
mypy-boto3-lambda==1.42.94
mypy-boto3-s3==1.42.94
mypy-boto3-schemas==1.42.3
mypy-boto3-secretsmanager==1.42.8
mypy-boto3-signer==1.42.7
mypy-boto3-sqs==1.42.3
mypy-boto3-stepfunctions==1.42.3
mypy-boto3-sts==1.42.91
mypy-boto3-xray==1.42.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
ollama==0.6.2
openai==2.33.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.2
pandas==3.0.2
pillow==12.2.0
pinecone==8.1.2
pinecone-plugin-assistant==3.0.3
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
propcache==0.4.1
pyarrow==24.0.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.14.0
pydantic_core==2.41.5
Pygments==2.20.0
pyOpenSSL==26.0.0
pypdf==6.10.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
ragas==0.4.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.4
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruff==0.15.12
s3transfer==0.16.1
safetensors==0.7.0
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sympy==1.14.0
tenacity==9.1.4
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.11.0
tqdm==4.67.3
transformers==5.7.0
typer==0.23.1
types-awscrt==0.31.3
types-requests==2.33.0.20260503
types-s3transfer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
watchdog==4.0.2
Werkzeug==3.1.8
wheel==0.47.0
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
